Changed database to work with new ideas
Instead of the approach i was taking, gonna make the process easier by storing the questions in the database and having the code pull the questions directly from the DB rather than what i was doing which would've taken way too much work. By the end the program should just be able to pull everything it needs directly from the DB rather than relying so much on the code.